Added BindingContextSheetContent property to simplify binding context of sheet content
